requirements.txt
albumentations==0.4.3
alembic==1.4.3
argon2-cffi==20.1.0
asciitree==0.3.3
async-generator==1.10
attrs==20.2.0
backcall==0.2.0
bleach==3.2.1
catboost==0.24.1
category-encoders==2.2.2
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cliff==3.4.0
cloudpickle==1.3.0
cmaes==0.6.1
cmd2==1.3.10
colorama==0.4.3
colorlog==4.2.1
cupy-cuda101==7.0.0
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
efficientnet-pytorch==0.7.0
entrypoints==0.3
fasteners==0.15
fastrlock==0.5
future==0.18.2
graphviz==0.14.1
idna==2.10
imageio==2.9.0
imgaug==0.2.6
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
joblib==0.14.1
json5==0.9.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyterlab==2.2.8
jupyterlab-pygments==0.1.2
jupyterlab-server==1.2.0
kiwisolver==1.2.0
l5kit==1.1.0
lightgbm==3.0.0
llvmlite==0.35.0
Mako==1.1.3
MarkupSafe==1.1.1
matplotlib==3.2.0
mistune==0.8.4
monotonic==1.5
mpi4py==3.0.3
munch==2.5.0
nbclient==0.5.0
nbconvert==6.0.6
nbformat==5.0.7
nest-asyncio==1.4.1
networkx==2.4
nose==1.3.7
notebook==6.1.4
numba==0.52.0
numcodecs==0.7.2
numpy==1.18.1
opencv-contrib-python-headless==4.4.0.44
opencv-python==4.2.0.32
opencv-python-headless==4.4.0.44
optuna==1.3.0
packaging==20.4
pandas==1.0.2
pandocfilters==1.4.2
parso==0.7.1
patsy==0.5.1
pbr==5.5.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
plotly==4.10.0
pretrainedmodels==0.7.4
prettytable==0.7.2
prometheus-client==0.8.0
prompt-toolkit==3.0.7
protobuf==3.13.0
PTable==0.9.2
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.7.1
pygobject==3.26.1
pymap3d==2.4.3
pyparsing==2.4.7
pyperclip==1.8.0
pyrsistent==0.17.3
python-apt==1.6.5+ubuntu0.3
python-dateutil==2.8.1
python-editor==1.0.4
pytorch-ignite==0.4.1
pytorch-pfn-extras==0.3.1
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==19.0.2
qtconsole==4.7.7
QtPy==1.9.0
requests==2.24.0
resnest==0.0.6b20200912
retrying==1.3.3
scikit-image==0.16.2
scikit-learn==0.22.2.post1
scipy==1.4.1
seaborn==0.9.0
segmentation-models-pytorch==0.1.0
Send2Trash==1.5.0
Shapely==1.7.1
six==1.15.0
SQLAlchemy==1.3.19
statsmodels==0.12.0
stevedore==3.2.2
terminado==0.9.1
testpath==0.4.4
timm==0.3.1
toml==0.10.0
torch==1.6.0+cu101
torchvision==0.7.0
tornado==6.0.4
tqdm==4.45.0
traitlets==5.0.4
transforms3d==0.3.1
urllib3==1.25.10
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
xgboost==1.2.0
zarr==2.4.0